# Results Comparer

This simple tool allows for easy comparison of provided benchmark results.

It can be used to compare:
* historical results (eg. before and after my changes)
* results for different OSes (eg. Windows vs Ubuntu)
* results for different CPU architectures (eg. x64 vs ARM64)
* results for different target frameworks (eg. .NET Core 2.1 vs 2.2)

All you need to provide is:
* `--base` - path to folder/file with baseline results
* `--diff` - path to folder/file with diff results
* `--treshold`  - threshold for Statistical Test. Examples: 5%, 10ms, 100ns, 1s

Optional arguments:
* `--top` - filter the diff to top/bottom `N` results

Sample: compare the results stored in `C:\results\windows` vs `C:\results\ubuntu` using `1%` threshold and print only TOP 10.

```cmd
dotnet run --base "C:\results\windows" --diff "C:\results\ubuntu" --treshold 1% --top 10
```

**Note**: the tool supports only `*full.json` results exported by BenchmarkDotNet. This exporter is enabled by default in this repository.

**Note**: if you have run your benchmarks for multiple jobs (eg. `-r netcoreapp2.1 netcoreapp2.2`) and you want to compare these historical results you can use `--merge` to point to the folder/file with such results.

## Reading the results

Sample results:

```log
SLOWER:
-30,35% BenchmarksGame.SpectralNorm_3.RunBench [can have several modes]
-5,45% BenchmarksGame.KNucleotide_9.RunBench [bimodal]
-4,27% BenchmarksGame.BinaryTrees_2.RunBench
-2,42% BenchmarksGame.MandelBrot_7.Bench(size: 4000, lineLength: 500, checksum: "C7-E6-66-43-66-73-F8-A8-D3-B4-D7-97-2F-FC-A1-D3")

FASTER:
3,59% BenchmarksGame.FannkuchRedux_5.RunBench(n: 10, expectedSum: 38)
```

Explanation:
* if there is no difference, the results are omitted
* if there is no match (we use full benchmark names to match the benchmarks), the results are omitted
* every line contains:
  * ratio (`1.0 - diffMedian / baseMedian`)
  * id (full benchmark name, the same we use in BenchView)
  * optional information about the modality of the benchmark