@brianrob 2018-12-04
@adamsitnik, will take a look.  Also, I suspect that @billwert will want to review as well, but he's OOF today.