@AndyAyersMS 2018-12-04
Looks like my arm64 files have some null entries:
```json
{
   "Title":"BenchmarksGame.BinaryTrees_2",
   "HostEnvironmentInfo":{
      "BenchmarkDotNetCaption":"BenchmarkDotNet",
      "BenchmarkDotNetVersion":"0.11.3.886-nightly",
      "OsVersion":"ubuntu 16.04",
      "ProcessorName":"Unknown processor",
      "PhysicalProcessorCount":null,
      "PhysicalCoreCount":null,
      "LogicalCoreCount":null,
      "RuntimeVersion":".NET Core 3.0.0-preview-27122-01 (CoreCLR 4.6.27121.03, CoreFX 4.7.18.57103)",
      "Architecture":"64bit",
      "HasAttachedDebugger":false,