@AndyAyersMS 2018-12-04
Thanks, it's working on my data now:
```
SLOWER:
-67880.50% BenchmarksGame.SpectralNorm_3.RunBench [can have several modes]
-8524.79% System.Memory.ReadOnlySpan.IndexOfString(input: "?", value: "?", comparisonType: InvariantCulture)
-2063.05% System.Memory.ReadOnlySpan.GetPinnableReference
-1638.45% System.Memory.ReadOnlySpan.StringAsSpan [bimodal]
-980.13% PerfLabTests.CastingPerf2.CastingPerf.ObjInt
-840.98% PerfLabTests.CastingPerf2.CastingPerf.ObjObjrefValueType
-839.70% PerfLabTests.CastingPerf2.CastingPerf.ObjScalarValueType
-809.60% PerfLabTests.LowLevelPerf.GenericClassWithSTringGenericInstanceMethod
-799.23% Functions.MathTests.PowSingleBenchmark
-768.68% System.Threading.Tests.Perf_Interlocked.Decrement_int

FASTER:
78.23% BenchmarksGame.ReverseComplement_6.RunBench
64.30% BenchmarksGame.MandelBrot_7.Bench(size: 4000, lineLength: 500, checksum: "C7-E6-66-43-66-73-F8-A8-D3-B4-D7-97-2F-FC-A1-D3") [can have several modes]
42.61% Burgers.Test1
41.78% Burgers.Test0
30.05% Benchstone.BenchI.Pi.Test
13.01% System.Memory.ReadOnlySpan.IndexOfString(input: "5555555555", value: "5", comparisonType: InvariantCulture)
11.28% System.Memory.ReadOnlySpan.IndexOfString(input: "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX", value: "x", comparisonType: InvariantCultureIgnoreCase)
```

Couple of things I'd like to see:
* ratios instead of percentages
* option to show base and diff values as well
* option for csv or other structured output so I can do further analysis on the distribution of ratios