@jorive 2018-12-04
Maybe we should have two more options:
`--format` output format: md, csv, etc.
`--comparison-output [ base/diff | diff/base |  percentage]` ?