@adamsitnik 2018-12-04
>  Is this designed for devs to just quickly iterate on their own or is this intended to be part of the reporting infrastructure of the performance automation system?

one thing is Windows vs Ubuntu or x64 vs ARM64

the other is: I run the benchmarks before applying any changes and save the results somewhere, then I apply the changes and save them to other location and compare the perf.