@adamsitnik 2018-12-04
> ratios instead of percentages

@AndyAyersMS would `--display ratio|%|delta` be OK?

> option to show base and diff values as well

ok, then I most probably need to introduce a table

> option for csv or other structured output so I can do further analysis on the distribution of ratios

@AndyAyersMS would sth like following simple JSON be enough?

```json
{
    "Benchmarks":[
    {
        "FullName": "BenchmarksGame.BinaryTrees_2.RunBench",
        "Base": [ 1, 2, 3, 4, 5],
        "Diff": [ 1, 2, 3, 4, 5]
    },
    {
        "FullName": "BenchmarksGame.BinaryTrees_5.RunBench",
        "Base": [ 1, 2, 3, 4, 5],
        "Diff": [ 1, 2, 3, 4, 5]
    } ]
}
```