@AndyAyersMS 2018-12-04
Ideally something that I can easily import into Excel or R ... I would think CSV would be the simplest.