@jorive 2018-12-05
@adamsitnik Should we add `*.sln` to `.gitignore`?