@adamsitnik 2018-12-05
> Ideally something that I can easily import into Excel or R ... I would think CSV would be the simplest.

@AndyAyersMS would sth like this be OK?

```csv
Conclusion;Id;Values
Base;System.Memory.Span.SomeMethod;0.01;0.02;0.03;
Slower;System.Memory.Span.SomeMethod;0.1;0.2;0.3;
```