@billwert 2018-12-05
> one thing is Windows vs Ubuntu or x64 vs ARM64
> the other is: I run the benchmarks before applying any changes and save the results somewhere, then I apply the changes and save them to other location and compare the perf.

@adamsitnik sure, makes sense. I'm more curious about whether or not this is something we'd want to incorporate into our automation strategy in general (think reporting), because modularizing it would be nice in that case. We can cross this bridge later however.

One other question: Why is this stand alone and not a mode in Benchmark.NET?