@jorive 2018-12-05
> One other question: Why is this stand alone and not a mode in Benchmark.NET?

@billwert personally, I think it's better to have tools that build on top of each other and that perform one task very well, instead of a monolithic tool that becomes too bloated and complex to maintain (the linux way some would say?). This way you could think of benchmark/runner/reporter