@adamsitnik 2018-12-05
Sample results:

| Slower                                                                                               | diff/base | Base Median (ns) | Diff Median (ns) | Modality              |
| ---------------------------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| ---------------------- |
| Burgers.Test1                                                                                        |     11.93 |     267235588.00 |    3187167559.50 |                       |
| System.Globalization.Tests.Perf_CompareInfo.IsSortable(text: "Hello Worldbbbbbbbbbbbbbbbbbbbbbbbbbbb |     11.63 |          1115.77 |         12974.74 |                       |
| System.Globalization.Tests.Perf_CompareInfo.IsSortable(text: "More Test's")                          |     11.24 |            68.36 |           768.48 |                       |
| System.Globalization.Tests.Perf_CompareInfo.IsSortable(text: "Exhibit A")                            |     11.10 |            56.58 |           628.19 |                       |
| System.Globalization.Tests.Perf_CompareInfo.IsSortable(text: "TestFooBA`RnotsolongTELLme")           |     11.00 |           163.97 |          1804.41 |                       |
| System.Numerics.Tests.Perf_Vector4.DotBenchmark                                                      |     10.13 |             1.50 |            15.16 |                       |
| System.Globalization.Tests.Perf_CompareInfo.IsSortable(text: "foo")                                  |      9.86 |            21.21 |           209.16 |                       |
| System.Globalization.Tests.Perf_CompareInfo.IsSortable(text: "$")                                    |      7.57 |             9.46 |            71.62 |                       |
| System.Globalization.Tests.Perf_CompareInfo.IsSortable(text: "?")                                    |      7.45 |             9.45 |            70.45 |                       |
| Burgers.Test0                                                                                        |      7.43 |     440849666.00 |    3275757908.00 |                       |
| System.Numerics.Tests.Perf_Vector4.AddFunctionBenchmark                                              |      6.21 |             1.82 |            11.30 |                       |
| System.Numerics.Tests.Perf_Vector4.AddOperatorBenchmark                                              |      6.07 |             1.82 |            11.07 |                       |
| FractalPerf.Launch.Test                                                                              |      5.75 |     174073993.00 |    1000937680.50 |                       |
| Benchstone.BenchI.Array2.Test                                                                        |      4.80 |     501337097.00 |    2405957355.00 |                       |
| Benchstone.BenchI.BubbleSort2.Test                                                                   |      2.94 |      35372702.64 |     103894135.00 | can have several modes|
| System.Tests.Perf_Double.ToStringWithFormat(format: "R", number: 1.79769313486232E+308, innerIterati |      2.92 |           372.24 |          1088.56 |                       |
| System.Tests.Perf_Double.ToStringWithFormat(format: "R", number: -1.79769313486232E+308, innerIterat |      2.90 |           375.55 |          1088.52 |                       |
| SciMark2.kernel.benchSparseMult                                                                      |      2.68 |     767957871.00 |    2054656061.50 |                       |
| Burgers.Test2                                                                                        |      2.58 |     267264201.00 |     689646694.00 |                       |
| Benchstone.BenchI.IniArray.Test                                                                      |      2.54 |     112425878.50 |     285823839.00 |                       |
| Benchstone.BenchF.BenchMrk.Test                                                                      |      2.43 |     184005617.00 |     446363534.50 |                       |
| System.Net.Http.Tests.SocketsHttpHandlerPerfTest.Get(ssl: True, chunkedResponse: False, responseLeng |      2.41 |       1620154.00 |       3910963.00 |                       |
| System.Net.Http.Tests.SocketsHttpHandlerPerfTest.Get(ssl: True, chunkedResponse: True, responseLengt |      2.38 |       1664260.00 |       3955205.50 |                       |
| SeekUnroll.Test(boxedIndex: 27)                                                                      |      2.32 |    2416673122.00 |    5597261928.00 |                       |
| BenchmarksGame.FannkuchRedux_2.RunBench(n: 10, expectedSum: 73196)                                   |      2.20 |     171421112.00 |     377649577.00 |                       |
| System.Diagnostics.Perf_Process.StartAndKillDotNetVersion                                            |      2.19 |        723188.79 |       1584806.65 |                       |
| System.Tests.Perf_Array.ArrayCreate1D                                                                |      2.18 |          1066.30 |          2324.38 |                       |
| System.Tests.Perf_Array.ArrayCreate3D                                                                |      2.17 |          1191.00 |          2580.83 |                       |
| Benchstone.BenchF.Lorenz.Test                                                                        |      2.13 |     273039368.00 |     581460525.00 |                       |
| SeekUnroll.Test(boxedIndex: 19)                                                                      |      2.11 |    2368672138.00 |    5004807345.00 |                       |
| System.Collections.CtorGivenSizeNonGeneric.Hashtable(Size: 512)                                      |      2.11 |          1249.62 |          2632.27 |                       |
| System.Tests.Perf_Array.ArrayCreate2D                                                                |      2.08 |          1192.54 |          2483.94 |                       |
| SeekUnroll.Test(boxedIndex: 11)                                                                      |      2.04 |    2453818595.00 |    5004047922.00 |                       |
| System.Tests.Perf_StringBuilder.StringBuilderAppend                                                  |      1.94 |        196651.17 |        381427.68 |                       |
| BenchmarksGame.Mandelbrot_2.Bench(width: 4000, checksum: "C7-E6-66-43-66-73-F8-A8-D3-B4-D7-97-2F-FC- |      1.93 |    1353477955.00 |    2609159425.00 |                       |
| Benchstone.BenchF.Romber.Test                                                                        |      1.82 |     167909839.00 |     305344466.50 |                       |
| System.Memory.ReadOnlySpan.StringAsSpan                                                              |      1.71 |             6.70 |            11.44 |                       |

| Faster                                                                                               | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ---------------------------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Tests.Perf_String.Contains(size: 1000)                                                        |     16.01 |           370.07 |            23.12 |         |
| System.Memory.ReadOnlySpan.IndexOfString(input: "Hello Worldbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbareally |      8.42 |           293.70 |            34.89 |         |
| System.Globalization.Tests.Perf_CompareInfo.IndexOf(culture: en-US, source: "Hello Worldbbbbbbbbbbbb |      7.58 |           260.58 |            34.36 |         |
| XmlDocumentTests.XmlDocumentTests.Perf_XmlDocument.Create                                            |      4.15 |          2446.32 |           589.30 |         |
| System.Tests.Perf_String.IndexOf(options: Ordinal)                                                   |      4.09 |         15972.11 |          3901.83 |         |
| System.Memory.ReadOnlySpan.IndexOfString(input: "Hello Worldbbbbbbbbbbbbbbcbbbbbbbbbbbbbbbbbbba!", v |      3.90 |           108.62 |            27.82 |         |
| System.Tests.Perf_Boolean.Parse_str                                                                  |      3.20 |            56.32 |            17.61 |         |
| System.Globalization.Tests.Perf_CompareInfo.IsPrefix(culture: en-US, source: "StrIng", prefix: "str" |      3.14 |            32.82 |            10.46 |         |
| System.Tests.Perf_String.Contains(text: "This is a very nice sentence", value: "bad", comparisonType |      2.98 |            45.03 |            15.12 |         |
| System.Globalization.Tests.Perf_CompareInfo.IndexOf(culture: es-ES, source: "Hello Worldbbbbbbbbbbbb |      2.88 |            80.92 |            28.14 |         |
| System.Globalization.Tests.Perf_CompareInfo.IsPrefix(culture: ja-JP, source: "XXXXXXXXXXXXXXXXXXXXXX |      2.72 |            29.56 |            10.87 |         |
| System.Globalization.Tests.Perf_CompareInfo.IsPrefix(culture: , source: "XXXXXXXXXXXXXXXXXXXXXXXXXXX |      2.64 |            29.05 |            11.02 |         |
| System.Net.Primitives.Tests.CredentialCacheTests.GetCredential_HostPort(host: "name5", hostPortCount |      2.45 |           230.66 |            94.31 |         |
| System.Tests.Perf_String.Compare(strings: ["Thé quick brown fox", "Thé quick BROWN fox"], comparison |      2.41 |            15.87 |             6.58 |         |
| System.Tests.Perf_String.GetHashCode(s: "")                                                          |      2.38 |             5.57 |             2.34 |         |
| System.Net.Primitives.Tests.CredentialCacheTests.GetCredential_HostPort(host: "notfound", hostPortCo |      2.36 |           183.45 |            77.86 |         |
| System.Tests.Perf_UInt64.Parse(value: "18446744073709551615")                                        |      2.34 |           159.40 |            68.21 |         |
| System.Tests.Perf_String.Compare(strings: ["The quick brown fox", "THE QUICK BROWN FOX"], comparison |      2.28 |            13.43 |             5.90 |         |
| System.Tests.Perf_Int64.Parse(value: "9223372036854775807")                                          |      2.21 |           150.47 |            68.17 |         |
| System.Tests.Perf_Int64.Parse(value: "-9223372036854775808")                                         |      2.13 |           151.89 |            71.18 |         |
| System.Globalization.Tests.Perf_CompareInfo.IsPrefix(culture: , source: "5555555555", prefix: "AAAAA |      2.06 |            16.87 |             8.20 |         |
| System.Globalization.Tests.Perf_CompareInfo.IsPrefix(culture: es-ES, source: "?????????????????????? |      2.05 |            16.65 |             8.13 |         |
| System.Net.Primitives.Tests.IPAddressPerformanceTests.TryFormat(address: 143.24.20.36)               |      2.02 |            72.92 |            36.01 |         |
| System.Tests.Perf_UInt32.Parse(value: "4294967295")                                                  |      1.95 |           109.66 |            56.11 |         |
| System.Globalization.Tests.Perf_CompareInfo.IsPrefix(culture: , source: "foobardzsdzs", prefix: "Foo |      1.90 |            50.68 |            26.73 |         |
| System.Tests.Perf_String.ToLowerInvariant(s: "test")                                                 |      1.89 |            22.71 |            12.04 |         |
| System.Tests.Perf_Guid.ctor_str                                                                      |      1.88 |           328.81 |           174.58 |         |
| System.Tests.Perf_String.Contains(size: 100)                                                         |      1.88 |           192.41 |           102.19 |         |
| System.Tests.Perf_String.GetHashCode(s: "TeSt!")                                                     |      1.87 |             6.76 |             3.61 |         |
| System.Tests.Perf_Int32.Parse(value: "-2147483648")                                                  |      1.86 |           110.69 |            59.52 |         |
| System.Globalization.Tests.Perf_CompareInfo.IsPrefix(culture: es-ES, source: "Hello Worldbbbbbbbbbbb |      1.84 |            21.31 |            11.55 |         |
| System.Tests.Perf_Int32.Parse(value: "2147483647")                                                   |      1.83 |           108.12 |            59.12 |         |
| PerfLabTests.LowLevelPerf.GenericGenericMethod                                                       |      1.81 |        323704.26 |        178350.41 |         |
| System.Tests.Perf_UInt64.Parse(value: "12345")                                                       |      1.79 |            91.38 |            51.08 |         |
| LinqBenchmarks.Where01LinqQueryX                                                                     |      1.78 |     692132246.00 |     388216530.50 |         |
| System.Tests.Perf_Int64.Parse(value: "12345")                                                        |      1.78 |            90.71 |            50.90 |         |
| System.Memory.ReadOnlySpan.IndexOfString(input: "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX |      1.78 |            62.91 |            35.37 |         |
| LinqBenchmarks.Where01LinqMethodX                                                                    |      1.78 |     692317830.00 |     389405790.00 |         |
| System.Globalization.Tests.Perf_CompareInfo.IsSuffix(culture: es-ES, source: "?????????????????????? |      1.75 |            22.64 |            12.91 |         |
| System.Memory.ReadOnlySpan.IndexOfString(input: "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX |      1.75 |            40.40 |            23.05 |         |
| System.Memory.ReadOnlySpan.IndexOfString(input: "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx |      1.75 |            40.38 |            23.06 |         |
| System.Memory.ReadOnlySpan.IndexOfString(input: "??????????????????????????????????????????????????? |      1.73 |            39.82 |            23.06 |         |
| LinqBenchmarks.Where01LinqMethodNestedX                                                              |      1.72 |     777704640.00 |     452272892.00 |         |
| PerfLabTests.LowLevelPerf.EmptyStaticFunction5Arg                                                    |      1.72 |       3112353.87 |       1814303.51 |         |
| System.Tests.Perf_Int32.Parse(value: "12345")                                                        |      1.71 |            87.02 |            51.00 |         |
| System.Net.Primitives.Tests.IPAddressPerformanceTests.GetAddressBytes(address: 143.24.20.36)         |      1.70 |            19.81 |            11.63 |         |
| System.Net.Http.Tests.SocketsHttpHandlerPerfTest.Get(ssl: False, chunkedResponse: False, responseLen |      1.69 |        205232.42 |        121376.53 |         |